Restructure code by functionality
