+++ src/common/buttons/DeleteButton.stories.js
