+++ src/common/toggleSwitch/ToggleSwitch.js
