+++ src/common/toggleSwitch/ToggleSwitch.test.js
