+++ src/pages/AddTransactionPage.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
-import TransactionFormInput from '../form/TransactionFormInput'
-import ToggleSwitch from '../toggleSwitch/ToggleSwitch'
-import ReturnToOverviewButton from '../buttons/ReturnToOverviewButton'
+import TransactionFormInput from '../transactionComponents/form/TransactionFormInput'
+import ToggleSwitch from '../common/toggleSwitch/ToggleSwitch'
+import ReturnToOverviewButton from '../common/buttons/ReturnToOverviewButton'
 import styled from 'styled-components/macro'
 import { postNewTransactionEntry } from '../services/transactionServices'
